Refactor auth components and remove register subcomponents
Moved login-related components out of the 'auth' directory to a top-level 'login' directory for both cxc and web apps. Deleted unused AccountInfo and AdditionalInfo components from the register flow. Updated imports in affected pages to reflect the new component locations.
